Fixed warnings on this branch
