+++ lib/views/widgets/expandable/expanded_action_button.dart
@@ -1,0 +1,41 @@
+import 'dart:math' as math;
+
+import 'package:flutter/material.dart';
+
+@immutable
+class ExpandedActionButton extends StatelessWidget {
+  const ExpandedActionButton({
+    super.key,
+    required this.directionInDegrees,
+    required this.maxDistance,
+    required this.progress,
+    required this.child,
+  });
+
+  final double directionInDegrees;
+  final double maxDistance;
+  final Animation<double> progress;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedBuilder(
+      animation: progress,
+      builder: (context, child) {
+        final offset = Offset.fromDirection(
+          directionInDegrees * (math.pi / 180.0),
+          progress.value * maxDistance,
+        );
+        return Positioned(
+          right: 4.0 + offset.dx,
+          bottom: 4.0 + offset.dy,
+          child: Transform.rotate(
+            angle: (1.0 - progress.value) * math.pi / 2,
+            child: child!,
+          ),
+        );
+      },
+      child: FadeTransition(opacity: progress, child: child),
+    );
+  }
+}
